Move getHistoryById to PositionRepository
