+++ packages/backend/test/peripherals/database/StateUpdateRepository.test.ts
@@ -45,71 +45,6 @@ describe(StateUpdateRepository.name, () => {
     })
   })
 
-  it('gets position by id', async () => {
-    const positionId = 12345n
-
-    await repository.add({
-      stateUpdate: {
-        id: 1,
-        blockNumber: 1,
-        rootHash: PedersenHash.fake(),
-        factHash: Hash256.fake(),
-        timestamp: Timestamp(0),
-      },
-      positions: [
-        {
-          publicKey: StarkKey.fake('1'),
-          positionId,
-          collateralBalance: 0n,
-          balances: [{ assetId: AssetId('ETH-9'), balance: 20n }],
-        },
-      ],
-      prices: [{ assetId: AssetId('ETH-9'), price: 20n }],
-    })
-
-    await repository.add({
-      stateUpdate: {
-        id: 2,
-        blockNumber: 2,
-        rootHash: PedersenHash.fake(),
-        factHash: Hash256.fake(),
-        timestamp: Timestamp(0),
-      },
-      positions: [
-        {
-          publicKey: StarkKey.fake('1'),
-          positionId,
-          collateralBalance: 0n,
-          balances: [{ assetId: AssetId('BTC-10'), balance: 40n }],
-        },
-      ],
-      prices: [{ assetId: AssetId('BTC-10'), price: 40n }],
-    })
-
-    const position = await repository.getPositionHistoryById(positionId)
-
-    expect(position).toEqual([
-      {
-        stateUpdateId: 2,
-        publicKey: StarkKey.fake('1'),
-        positionId,
-        collateralBalance: 0n,
-        balances: [{ assetId: AssetId('BTC-10'), balance: 40n }],
-        prices: [{ assetId: AssetId('BTC-10'), price: 40n }],
-        timestamp: Timestamp(0),
-      },
-      {
-        stateUpdateId: 1,
-        publicKey: StarkKey.fake('1'),
-        positionId,
-        collateralBalance: 0n,
-        balances: [{ assetId: AssetId('ETH-9'), balance: 20n }],
-        prices: [{ assetId: AssetId('ETH-9'), price: 20n }],
-        timestamp: Timestamp(0),
-      },
-    ])
-  })
-
   it('gets state update id by root hash', async () => {
     const stateRootId = 1
     const positionId = 12345n
@@ -236,7 +171,7 @@ describe(StateUpdateRepository.name, () => {
   })
 
   it('gets last state update by block number', async () => {
-    let last = await repository.getLast()
+    let last = await repository.findLast()
 
     expect(last).toEqual(undefined)
 
@@ -262,7 +197,7 @@ describe(StateUpdateRepository.name, () => {
     }
     await repository.add({ stateUpdate, positions: [], prices: [] })
 
-    last = await repository.getLast()
+    last = await repository.findLast()
 
     expect(last).toEqual(stateUpdate)
   })
